% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgboost.R
\name{as_sgb_params}
\alias{as_sgb_params}
\alias{sgb_params}
\title{Survival Parameters for \link[xgboost:xgb.train]{xgboost}}
\usage{
as_sgb_params(params)

sgb_params(...)
}
\arguments{
\item{params}{a list of tuning parameters for an \code{xgboost} model.}

\item{...}{name-value pairs (e.g., maxdepth = 2) of tuning parameters
and tuning parameter values.}
}
\value{
a modified list with values for \code{objective} and \code{eval_metric}
}
\description{
Survival Parameters for \link[xgboost:xgb.train]{xgboost}
}
\examples{

sgb_params()

as_sgb_params(list(max_depth=1))

}
