% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgboost.R
\name{as_tibble.sgb_data}
\alias{as_tibble.sgb_data}
\title{Coerce \code{sgb_data} to Data Frame.}
\usage{
as_tibble.sgb_data(x, ...)
}
\arguments{
\item{x}{an \code{sgb_data} object}

\item{...}{Currently not used}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Coerce \code{sgb_data} to Data Frame.
}
\examples{

library(tibble)

df = data.frame(time=c(1,2,3), status = c(0,0,1), x = c(2,2,1))

df_sgb <- as_sgb_data(df, status = status, time = time)

tbl_sgb <- as_tibble(df_sgb)

tbl_sgb

}
