% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_cats.R
\name{cat_spread}
\alias{cat_spread}
\alias{cat_gather}
\title{Herding Cats!}
\usage{
cat_spread(data, ...)

cat_gather(data, factor_levels)
}
\arguments{
\item{data}{data with categorical variables (i.e., factors)
that need to be spread or gathered.}

\item{...}{Arguments passed on to \code{mltools::one_hot}
\describe{
  \item{sparsifyNAs}{Should NAs be converted to 0s?}
  \item{naCols}{Should columns be generated to indicate the present of NAs? Will only apply to factor columns with at least one NA}
  \item{dropCols}{Should the resulting data.table exclude the original columns which are one-hot-encoded?}
}}

\item{factor_levels}{This parameter is only relevant for
\code{cat_gather}. A named list of factor levels, with each
name corresponding to the column in the data that the
factor levels describe.}
}
\value{
a \link[tibble:tibble-package]{tibble} with categorical variables
herded as you like.
}
\description{
\code{xgboost} (and many other) modeling functions
expect matrix input with factor levels one-hot encoded.

\code{cat_spread} will one-hot encode any factor or character variable
in \code{data} and return a one-hot encoded \code{tibble}. Alternatively,
\code{cat_gather} will apply the inverse operation and convert one-hot
encoded columns back into factors.
}
\examples{

df <- data.frame(x = rep(letters[1:2], 50), y = 1:100)

one_hot_df <- cat_spread(df)
cat_gather(one_hot_df, factor_levels = list(x=c('a','b')))


}
