% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_cats.R
\name{cat_transfer}
\alias{cat_transfer}
\title{transfer factor levels}
\usage{
cat_transfer(to, from)
}
\arguments{
\item{to}{the data that factor levels are transferred to}

\item{from}{the data that factor levels are transferred from}
}
\description{
take the factor levels in training data
and copy them over to testing data. This is an important
pre-processing step for data splits that may have
different factor levels in training and testing sets.
}
\note{
\code{to} and \code{from} must have the same factor columns. For example,
if \code{to} has a factor named \code{A} and \code{from} does not have a factor
of the same name, the function will stop and tell you which
factor variables are missing.
}
