% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{lbl_survival}
\alias{lbl_survival}
\alias{lbl_multi}
\alias{lbl_binary}
\title{Labels for \link[xgboost:xgb.train]{xgboost}}
\usage{
lbl_survival(time, status)

lbl_multi(response)

lbl_binary(response)
}
\arguments{
\item{time}{a numeric vector with time-to-event values.}

\item{status}{a numeric vector of 0s and 1s, where 0 indicates
censoring and 1 indicates that the event occurred.}

\item{response}{a numeric, factor, or logical vector containing
label values.}
}
\description{
\code{xgboost} functions expect a certain type of label
vector for certain problem types. For survival models, \code{xgboost}
labels should be time-to-event values with negative times
indicating censored observations and postive times indicating
events. For binary classification problems, \code{xgboost} expects
a vector of 0's and 1's with 1 indicating an event and 0 indicating
no event. For K-class classification problems, \code{xgboost} expects
a numeric vector with values of 0, 1, ..., and K-1. The set of
functions described here work to make label creation a little
less tedious.
}
\note{
\code{xgboost} doesn't expect any special characteristics for
continous outcomes, so we did not develop a function to
handle them.
}
\examples{

lbl_survival(
  time = c(1,2,3),
  status = c(1,0,1)
)

lbl_binary(
  response = factor(x = c("A","A","B"))
)

lbl_multi(
  response = factor(x = c("A","C","B"))
)

}
