% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{sgb_bhaz}
\alias{sgb_bhaz}
\title{Baseline Hazard}
\usage{
sgb_bhaz(sgb_booster, eval_times = NULL, smooth = FALSE,
  cumulative = TRUE)
}
\arguments{
\item{sgb_booster}{an object of class \code{sgb_booster} (see \link{sgb_fit}).}

\item{eval_times}{Values at which the baseline hazard will
be evaluated.}

\item{smooth}{If \code{TRUE} \code{sgb_bhaz} will smooth the estimated
baseline hazard using Friedman's super smoother
\link[stats:supsmu]{supsmu}.}

\item{cumulative}{If \code{TRUE} the cumulative survival function
will be computed.}
}
\value{
A vector with a baseline hazard value corresponding to each
time in \code{eval_times}  If cumulative is \code{TRUE}, the returned vector
evaluates the cumulative hazard function at those values.
}
\description{
This function is a wrapper for the
\code{\link[gbm:basehaz.gbm]{gbm::basehaz.gbm()}}. The function
computes the Breslow estimator of the baseline
hazard function for a proportional hazard
regression model.
}
