% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgboost.R
\name{sgb_data}
\alias{sgb_data}
\alias{as_sgb_data}
\title{Create data for \code{sgboost}}
\usage{
sgb_data(data, label)

as_sgb_data(data, status, time)
}
\arguments{
\item{data}{the data containing predictor variables and a label column}

\item{label}{a numeric vector based on \code{time} and \code{status}
values. Time values should be less than zero for censored
observations, and greater than zero for non-censored
observations.}

\item{status}{a numeric vector indicating status at a given time.
Normally, 0 indicates no event and 1 indicates an event occurred.}

\item{time}{a numeric vector of follow-up time values.}
}
\value{
an object of class \code{sgb_data} with components:
\itemize{
\item \code{data}: a matrix with columns representing predictor variables
\item \code{label}: a numeric vector representing time until event. Negative
times indicate that an event did not occur, but the observation
was censored at the absolute value of the given time. Positive times
indicate the time of the event.
}
}
\description{
\link[xgboost:xgb.train]{xgboost} operates using a data structure
called \code{xgb.DMatrix}. The \code{xgboost} functions can internally create
these data structures if they are given a matrix with columns giving
predictor variables and a vector representing the label. For survival
analysis, the label vector is a combination of time/status values
(see \link{sgb_label}). This function automates the creation of a label
vector and creates a list of components that are easily plugged in
to \code{xgboost} functions.
}
\examples{

df = data.frame(time=c(1,2,3), status = c(0,0,1), x = c(2,2,1))

as_sgb_data(df, status = status, time = time)

sgb_data(df, label = sgb_label(df$time, df$status))

}
