% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgboost.R
\name{sgb_label}
\alias{sgb_label}
\title{Labels for \link[xgboost:xgb.train]{xgboost}}
\usage{
sgb_label(time, status)
}
\arguments{
\item{time}{a numeric vector with time-to-event values.}

\item{status}{a numeric vector of 0s and 1s, where 0 indicates
censoring and 1 indicates that the event occurred.}
}
\description{
\code{xgboost} functions expect a certain type of label
vector for certain problem types. For survival models, \code{xgboost}
labels should be time-to-event values with negative times
indicating censored observations and postive times indicating
events.
}
\examples{

sgb_label(
  time = c(1,2,3),
  status = c(1,0,1)
)

}
