% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{xgb_folds}
\alias{xgb_folds}
\title{Folds for \link[xgboost:xgb.train]{xgboost}}
\usage{
xgb_folds(data, nfolds, strata = NULL)
}
\arguments{
\item{data}{a \link{data.frame} or \link[tibble:tibble-package]{tibble} that
will be used to generate folds.}

\item{nfolds}{(integer) the number of folds}

\item{strata}{A variable that is used to conduct stratified sampling
to create the folds. This could be a single character value or a
variable name that corresponds to a variable in \code{data}.}
}
\description{
When calling \link[xgboost:xgb.cv]{xgb.cv}, analysts
may wish to manually specify \code{folds} rather than \code{nfolds}.
This function generates an object that can be given as
an input for \code{folds} and additionally lets the analyst
specify stratifying variables to create balanced folds.
}
\examples{

df <- data.frame(a=1:100, b=100:1, c=rep(letters[1:2],50))

xgb_folds(data = df, nfolds = 10)
xgb_folds(data = df, nfolds = 100)
xgb_folds(data = df, nfolds = 10, strata = c)
}
