% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.surv-package.R
\docType{package}
\name{xgboost.surv-package}
\alias{xgboost.surv}
\alias{xgboost.surv-package}
\title{xgboost.surv: Extreme Gradient Boosting for Survival Analysis}
\description{
Gradient boosted decision trees are an excellent
  machine learning algorithm to develop prediction functions. 
  For analyses with censored outcomes, there are a number of 
  ways to leverage the popular and efficient R package 'xgboost'.
  The xgboost.surv package provides a framework to help you 
  engage with these types of risk prediction analyses using `xgboost`.
}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@uab.edu} (0000-0001-7399-2299)

}
\keyword{internal}
